Refactor content.js to support multiple versions of web.telegram.org
This commit refactors the content.js file to support multiple versions of web.telegram.org. It updates the `getDomElementsForService` function to handle both version "A" and version "K" of the website. The changes ensure that the correct DOM elements are selected based on the version of the website being used.

Additionally, the commit improves the error handling in the `decryptText` function by checking for the presence of the encryption string in the correct format.